+++ scripts/eateries/Eateries.js
@@ -1,4 +1,4 @@
-export const eateries = (eateriesObj) => {
+export const Eateries = (eateriesObj) => {
   return `
   <h2 class="eateriesName">Eateries Name: ${eateriesObj.businessName}</h2>
   `;
