+++ scripts/main.js
@@ -5,7 +5,8 @@ import { ParkSelect } from "./parks/ParkSelect.js";
 import { getEateries, useEateries } from "./eateries/EateryProvider.js";
 import { EateriesSelect } from "./eateries/EateriesSelect.js";
 import { AttractionSelect } from "./attractions/AttractionsSelect.js";
-import "./Preview/Preview.js";
+import "./parks/ParkPreview.js";
+import "./eateries/EateriesPreview.js";
 
 useParks();
 getParks();
@@ -13,4 +14,4 @@ ParkSelect();
 getEateries();
 useEateries();
 EateriesSelect();
-AttractionSelect()
+AttractionSelect();
